[MMU3] Updated MMU3-12x `selector_positions` parameter.
Previously, I set the selector values by matching the lower notches on the selector. These happens to be slightly off compared to the notches on the upper part of the selector. After this update, the selector is now exactly where it needs to be on all 12 positions. This should eliminate some filament load issues that I had where I solved by wiggling the selector back and forth, obviously caused by misalignments.